requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
loguru==0.7.0
mysqlclient==2.2.0
PyJWT==2.8.0
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
schedule==1.2.0
sqlparse==0.4.4
telebot==0.0.5
tzdata==2023.3
urllib3==2.0.4
win32-setctime==1.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
idna==3.4
loguru==0.7.0
redis==4.6.0
requests==2.31.0
schedule==1.2.0
urllib3==2.0.4
win32-setctime==1.1.0
